// frontend/src/app/core/api/generated/models/incident-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface IncidentResponse {
  assetId?: string;
  description?: string;
  id?: string;
  location?: string;
  operationalSignals?: Array<string>;
  priority?: 'P1' | 'P2' | 'P3' | 'P4';
  reportedAt?: string;
  severity?: 'Low' | 'Medium' | 'High' | 'Critical';
  status?: 'Open' | 'Acknowledged' | 'In Progress' | 'Resolved';
  title?: string;
  version?: number;
}